@@ -2,13 +2,13 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <StartupObject>TabulateSmarterTestContentPackage.Program</StartupObject>
     <RuntimeIdentifiers>win-x64;win-x86</RuntimeIdentifiers>
     <NeutralLanguage />
-    <Version>3.4.5</Version>
-    <AssemblyVersion>3.4.5.0</AssemblyVersion>
-    <FileVersion>3.4.5.0</FileVersion>
+    <Version>3.4.5.9</Version>
+    <AssemblyVersion>3.4.5.9</AssemblyVersion>
+    <FileVersion>3.4.5.9</FileVersion>
     <Authors>Smarter Balanced Assessment Consortium</Authors>
     <Company>Smarter Balanced Assessment Consortium</Company>
     <Description>Validates and tabulates the contents of a Smarter Balanced Test Content Package</Description>
