@@ -18,16 +18,19 @@ namespace TabulateSmarterTestContentPackage.Extractors
             writer.WriteLine("<html><head><style>table {border-collapse: collapse;} table, th, td {border: 1px solid black;}</style></head><body><table>");
             writer.WriteLine("<tr><th>Scorepoint</th><th>Name</th><th>Content</th></tr>");
 
+            int contentLength = 0;
             foreach (XmlElement xmlRubric in rubrics)
             {
+                var text = xmlRubric.XpEvalE("val/text()");
+                contentLength += text.Length;
                 writer.WriteLine("<tr><td>{0}</td><td>{1}</td><td>{2}</td></tr>",
                     xmlRubric.XpEvalE("@scorepoint"),
                     xmlRubric.XpEvalE("name"),
-                    xmlRubric.XpEvalE("val/text()"));
+                    text);
             }
             writer.WriteLine("</table></body></html>");
 
-            return true;
+            return contentLength > 0;
         }
 
         static public bool ExtractRubric(XmlDocument xml, Stream stream)
